Small service refactor
